fix: date로 변환하는 함수 유틸파일로 분리
