+++ pennyway-client-iOS/pennyway-client-iOS/View/TabView/SpendingManagementView/MySpendingListView/MySpendingListView.swift
@@ -161,7 +161,7 @@ struct MySpendingListView: View {
         let grouped = Dictionary(grouping: spendingHistoryViewModel.dailyDetailSpendings, by: { String($0.spendAt.prefix(10)) })
         let sortedGroup = grouped.map { (key: $0.key, values: $0.value) }
             .sorted { group1, group2 -> Bool in
-                if let date1 = spendingHistoryViewModel.dateFromString(group1.key + " 00:00:00"), let date2 = spendingHistoryViewModel.dateFromString(group2.key + " 00:00:00") {
+                if let date1 = DateFormatterUtil.dateFromString(group1.key + " 00:00:00"), let date2 = DateFormatterUtil.dateFromString(group2.key + " 00:00:00") {
                     return date1 > date2
                 }
                 return false
